pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coral-collective"
dynamic = ["version"]
description = "AI Agent Orchestration Framework with 20+ Specialized Agents"
authors = [
    {name = "CoralCollective Contributors", email = "noreply@coral-collective.dev"}
]
license = {text = "AGPL-3.0"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["ai", "agents", "orchestration", "development", "automation", "claude", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "PyYAML>=6.0",
    "rich>=13.0.0",
    "pyperclip>=1.8.2",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "aiofiles>=23.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
memory = [
    "chromadb>=0.4.15",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.25.0",
    "torch>=2.0.0",
    "pandas>=2.0.0",
]
mcp = [
    "mcp>=0.3.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]
all = [
    "coral-collective[memory,mcp,dev,docs]",
]

[project.scripts]
coral = "coral_collective.cli.main:main"
coral-init = "coral_collective.cli.main:init_command"
coral-check = "coral_collective.utils.check:main"

[project.urls]
Homepage = "https://github.com/coral-collective/coral-collective"
Documentation = "https://coral-collective.dev/docs"
Repository = "https://github.com/coral-collective/coral-collective"
"Bug Tracker" = "https://github.com/coral-collective/coral-collective/issues"
Changelog = "https://github.com/coral-collective/coral-collective/blob/main/CHANGELOG.md"

[tool.setuptools.dynamic]
version = {attr = "coral_collective.__version__"}

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
coral_collective = [
    "agents/*.md",
    "agents/**/*.md",
    "config/*.yaml",
    "config/*.json",
    "templates/*.md",
    "templates/*.json",
    "mcp/configs/*.yaml",
    "memory/*.py",
]

# ================================
# DEVELOPMENT TOOLS CONFIGURATION
# ================================

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["ARG001"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.mypy]
python_version = "3.8"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
    "transformers.*",
    "torch.*",
    "mcp.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=coral_collective",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["coral_collective"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
known_first_party = ["coral_collective"]
known_third_party = ["pytest", "chromadb", "mcp"]

[tool.bandit]
exclude_dirs = ["tests", "venv", "env"]
skips = ["B101", "B601"]